+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # CHANGELOG
 
 ## [Unreleased]
-* List all expenses;
 * List all expenses by id;
 * Update an expense;
 * Logic delete an expense;
@@ -12,6 +11,11 @@
 * Add Configure swagger;
 * Add tags;
 * Tag expenses;
+
+## [v1.1.0]() - 2020-12-26
+## Added 
+* List all expenses paginated
+* Create test for list all expenses in service layer;
 
 ## [v1.0.0]() - 2020-12-23
 ## Added